* origin/devel: (31 commits)
  bug-fix/internal-issue-#591 (#9639)
  miscellaneous adjustments to pacify cppcheck (#9677)
  Bug fix/simon 19 08 07 (#9655)
  add a way to output the version in the tests (#9673)
  removed unused variable
  hide MMFiles-specific information when we don't need it (#9667)
  remove unused instance variable (#9662)
  Add missing includes for Windows builds. (#9657)
  Set jsunity output to gtest style (#9646)
  add the ability to have cluster special examples (#9613)
  Create TakeoverShardLeader job. (#9653)
  Aardvark: Add k Shortest Paths example graph to UI (#9491)
  [devel] Resign Leadership (#9427)
  get rid of _resolvedNames and _nameLock (#9649)
  Add constructors to fix cppcheck warnings. (#9648)
  Bug fix/optimizations 06 08 2019 (#9641)
  fix compile error
  Decentralize includes (#9623)
  add error to failure (#9637)
  Fast Controlled Leaderchange (#9608)
  ...

+++ arangod/RestHandler/RestBatchHandler.cpp
@@ -28,7 +28,9 @@
 #include "GeneralServer/GeneralServer.h"
 #include "GeneralServer/GeneralServerFeature.h"
 #include "GeneralServer/RestHandlerFactory.h"
+#include "Logger/LogMacros.h"
 #include "Logger/Logger.h"
+#include "Logger/LoggerStream.h"
 #include "Rest/HttpRequest.h"
 #include "Rest/HttpResponse.h"
 #include "Scheduler/SchedulerFeature.h"
@@ -117,7 +119,8 @@ void RestBatchHandler::processSubHandlerResult(RestHandler const& handler) {
     httpResponse->body().appendText(_boundary + "--");
 
     if (_errors > 0) {
-      httpResponse->setHeaderNC(StaticStrings::Errors, StringUtils::itoa(_errors));
+      httpResponse->setHeaderNC(StaticStrings::Errors,
+                                StringUtils::itoa(static_cast<uint64_t>(_errors)));
     }
     continueHandlerExecution();
   } else {
